backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-jose[cryptography]==3.3.0
pydantic-settings==2.7.1
python-dotenv==1.0.1
sqlmodel==0.0.22
httpx==0.28.1
alembic==1.14.1
psycopg2-binary==2.9.10
